COLEGIO/Colegio/requirements.txt
Django==2.0.6
django-adminlte2==0.3.0
django-registration-redux==2.2
djangorestframework==3.8.2
Pillow==5.1.0
psycopg2-binary==2.7.4
pytz==2018.4
